使用 size_t 存放 msg_createwindow::id

+++ src/ege.h
@@ -569,7 +569,7 @@ struct msg_createwindow {
 	LPCWSTR classname;
 	DWORD style;
 	DWORD exstyle;
-	unsigned id;
+	size_t id;
 	LPVOID param;
 };
